IG/IGL Generators and Related Functions: matching inputs

These are the psi, H, and kappa functions of the IG and IGL copula families, but with inputs needing to be vectors of the same length. These functions are called by the R functions of the same name, without the _vec suffix.

Usage

igl_gen_vec(x, alpha)

igl_gen_D_vec(x, alpha)

igl_gen_inv_vec(p, alpha)

igl_kappa_vec(x, alpha)

igl_kappa_D_vec(x, alpha)

igl_kappa_inv_vec(p, alpha)

interp_gen_vec(x, eta, alpha)

interp_gen_inv_vec(p, eta, alpha)

interp_kappa_vec(x, eta, alpha)

interp_kappa_inv_vec(p, eta, alpha)

Arguments

x

Function argument. Vector of non-negative values.

p

Function inverse argument. Vector of values between 0 and 1.

eta, alpha

Function parameters. Vector of positive values.

Note

If calling this function manually, make sure each input are vectors of a common length.

See also