IG Copula Family Functions

Functions related to the IG copula family, denoted by 'ig'.

Usage

pcondig21(v, u, theta, alpha)

qcondig21(p, u, theta, alpha)

qcondig(p, u, theta, alpha)

pcondig(v, u, theta, alpha)

pcondig12(u, v, theta, alpha)

qcondig12(p, v, theta, alpha)

dig(u, v, theta, alpha)

logdig(u, v, theta, alpha)

pig(u, v, theta, alpha)

rig(n, theta, alpha)

Arguments

u, v

Vectors of values between 0 and 1 representing values of the first and second copula variables.

theta

Parameter of the IG copula family. Vectorized; >0.

alpha

Parameter of the IG copula family. Vectorized; >0.

p

Vector of quantile levels between 0 and 1 to evaluate a quantile function at.

n

Positive integer. Number of observations to randomly draw.

Value

Numeric vector of length equal to the length of the input vector(s).

Note

Inputting two vectors greater than length 1 is allowed, if they're the same length. Also, qcondig21 and pcondig21 are the same as qcondig and pcondig – they're the distributions of variable 2 given 1.

Examples

u <- runif(10)
v <- runif(10)
pig(u, v, theta = 5, alpha = 1)
#>  [1] 0.079648326 0.162368477 0.026362046 0.074323196 0.004455655 0.122812692
#>  [7] 0.257811554 0.027490461 0.330950541 0.768651255
dig(u, v, theta = 2, alpha = 2)
#>  [1] 0.6531305 0.8742874 0.9792254 1.1742052 1.2351493 0.9896548 0.9901719
#>  [8] 1.0359737 0.8771307 1.6711461
logdig(u, v, theta = 2, alpha = 2)
#>  [1] -0.425978390 -0.134346098 -0.020993430  0.160591535  0.211191866
#>  [6] -0.010399079 -0.009876691  0.035341735 -0.131099211  0.513509688
pcondig21(v, u, theta = 3, alpha = 6)
#>  [1] 0.91720529 0.17493670 0.03424133 0.32060350 0.40474305 0.19566369
#>  [7] 0.40310912 0.06366146 0.38828096 0.96078162
qcondig21(v, u, theta = 3, alpha = 6)
#>  [1] 0.83368782 0.17494456 0.03424133 0.32016882 0.39998875 0.19567598
#>  [7] 0.40396967 0.06366146 0.38912428 0.98621554
pcondig12(u, v, theta = 3, alpha = 6)
#>  [1] 0.078299811 0.834355112 0.600760902 0.158437311 0.007645536 0.466513005
#>  [7] 0.501368832 0.289767791 0.734541432 0.704784075
qcondig12(u, v, theta = 3, alpha = 6)
#>  [1] 0.083206392 0.834310960 0.600760870 0.155984470 0.007161119 0.466273968
#>  [7] 0.494164675 0.289766698 0.731212224 0.824525886
rig(10, theta = 3, alpha = 3)
#> # A tibble: 10 × 2
#>         u      v
#>     <dbl>  <dbl>
#>  1 0.290  0.0312
#>  2 0.678  0.233 
#>  3 0.735  0.318 
#>  4 0.196  0.575 
#>  5 0.981  0.533 
#>  6 0.742  0.470 
#>  7 0.0514 0.600 
#>  8 0.530  0.944 
#>  9 0.696  0.185 
#> 10 0.689  0.224 

# log density available for extra precision
log(dig(0.1, 0.1, 2.5, 12.3)) == logdig(0.1, 0.1, 2.5, 12.3)
#> [1] FALSE