IG Copula Family Functions

Functions related to the IG copula family, denoted by 'ig'.

pcondig21(v, u, cpar)

qcondig21(tau, u, cpar)

qcondig(tau, u, cpar)

pcondig(v, u, cpar)

qcondig12(tau, v, cpar, mxiter = 80, eps = 1e-12, bd = 5)

pcondig12(u, v, cpar)

dig(u, v, cpar)

logdig(u, v, cpar)

pig(u, v, cpar)

rig(n, cpar)

Arguments

u, v
Vectors of values between 0 and 1 representing values of the first and second copula variables.

cpar
Vector of length 2 corresponding to the copula parameters theta>0 and alpha>0, respectively.

tau
Vector of quantile levels between 0 and 1 to evaluate a quantile function at.

mxiter
Maximum number of iterations to run the Newton-Raphson algorithm when computing inverse. Positive integer, default 20.

eps
The Newton-Raphson algorithm for computing an inverse will stop if the step size is less than this small number.

bd
The largest acceptable step size in the Newton-Raphson algorithm. Step size is reduced if it reaches this large.

n
Positive integer. Number of observations to randomly draw.

Value

Numeric vector of length equal to the length of the input vector(s).

Note

Inputting two vectors greater than length 1 is allowed, if they're the same length. Also, qcondig21 and pcondig21 are the same as qcondig and pcondig -- they're the distributions of variable 2 given 1.

Examples

set.seed(1)
u <- runif(10)
v <- runif(10)
pig(u, v, cpar = c(5, 1))

#>  [1] 0.07998850 0.09248525 0.46307012 0.36915934 0.18767583 0.47424038
#>  [7] 0.69702131 0.66064796 0.29172181 0.05846469

dig(u, v, cpar = c(2, 2))

#>  [1] 1.0949890 1.0328600 1.0258831 0.8253385 0.9488539 0.8400993 0.9415070
#>  [8] 1.2682942 0.9214325 0.8553771

logdig(u, v, cpar = c(2, 2))

#>  [1]  0.09074430  0.03233167  0.02555377 -0.19196166 -0.05250043 -0.17423516
#>  [7] -0.06027354  0.23767289 -0.08182576 -0.15621280

pcondig21(v, u, cpar = c(3, 6))

#>  [1] 0.2059765 0.1765546 0.6794343 0.3837089 0.7863289 0.4960070 0.7068927
#>  [8] 0.9873761 0.3796761 0.8153175

qcondig21(v, u, cpar = c(3, 6))

#>  [1] 0.2059726 0.1765589 0.6949935 0.3845008 0.7546750 0.4994238 0.7291069
#>  [8] 0.9947989 0.3803962 0.7442596

pcondig12(u, v, cpar = c(3, 6))

#>  [1] 0.26568654 0.37220495 0.59720796 0.90874810 0.22671056 0.90028096
#>  [7] 0.94829038 0.44152809 0.63118042 0.07042358

qcondig12(u, v, cpar = c(3, 6))

#>  [1] 0.26533079 0.37204284 0.54733005 0.90766428 0.17877750 0.89646254
#>  [7] 0.94080742 0.79626601 0.62703647 0.05417459

rig(10, cpar = c(3, 3))

#> # A tibble: 10 x 2
#>         u     v
#>     <dbl> <dbl>
#>  1 0.935  0.537
#>  2 0.212  0.547
#>  3 0.652  0.531
#>  4 0.126  0.179
#>  5 0.267  0.768
#>  6 0.386  0.652
#>  7 0.0134 0.669
#>  8 0.382  0.108
#>  9 0.870  0.809
#> 10 0.340  0.402