IG Copula Family Functions

Functions related to the IG copula family, denoted by 'ig'.

pcondig21(v, u, cpar)

qcondig21(tau, u, cpar)

qcondig(tau, u, cpar)

pcondig(v, u, cpar)

qcondig12(tau, v, cpar, mxiter = 80, eps = 1e-12, bd = 5)

pcondig12(u, v, cpar)

dig(u, v, cpar)

logdig(u, v, cpar)

pig(u, v, cpar)

rig(n, cpar)

Arguments

u, v
Vectors of values between 0 and 1 representing values of the first and second copula variables.

cpar
Vector of length 2 corresponding to the copula parameters theta>0 and alpha>0, respectively.

tau
Vector of quantile levels between 0 and 1 to evaluate a quantile function at.

mxiter
Maximum number of iterations to run the Newton-Raphson algorithm when computing inverse. Positive integer, default 20.

eps
The Newton-Raphson algorithm for computing an inverse will stop if the step size is less than this small number.

bd
The largest acceptable step size in the Newton-Raphson algorithm. Step size is reduced if it reaches this large.

n
Positive integer. Number of observations to randomly draw.

Value

Numeric vector of length equal to the length of the input vector(s).

Note

Inputting two vectors greater than length 1 is allowed, if they're the same length. Also, qcondig21 and pcondig21 are the same as qcondig and pcondig -- they're the distributions of variable 2 given 1.