IGL Copula Family Functions

Functions related to the IGL copula family, denoted by 'igl'.

qcondigl(p, u, alpha)

pcondigl(v, u, alpha)

qcondigl21(p, u, alpha)

pcondigl21(v, u, alpha)

pcondigl12(u, v, alpha)

qcondigl12(p, v, alpha)

digl(u, v, alpha)

pigl(u, v, alpha)

rigl(n, alpha)

logdigl(u, v, alpha)

Arguments

p
Vector of quantile levels between 0 and 1 to evaluate a quantile function at.

u, v
Vectors of values between 0 and 1 representing values of the first and second copula variables.

alpha
Single numeric >0; corresponds to parameter alpha in the IGL copula family.

n
Positive integer. Number of observations to randomly draw.

Value

Numeric vector of length equal to the length of the input vector(s).

Note

Inputting two vectors greater than length 1 is allowed, if they're the same length. Also, qcondigl21 and pcondigl21 are the same as qcondigl and pcondigl -- they are the distributions of variable 2 given 1.

Examples

set.seed(1)
u <- runif(10)
v <- runif(10)
pigl(u, v, alpha = 1)

#>  [1] 0.09036086 0.10358874 0.49893409 0.37976926 0.19726813 0.48995618
#>  [7] 0.71269122 0.66079779 0.31652482 0.06094790

digl(u, v, alpha = 2)

#>  [1] 1.573817e+00 1.351845e+00 1.628684e+00 1.317590e-01 2.054155e-01
#>  [6] 2.514377e-01 3.743734e-01 3.446458e-31 9.890313e-01 7.031873e-02

logdigl(u, v, alpha = 0.4)

#>  [1]   0.20792570   0.14927593   0.11282773  -0.53704402  -0.16615410
#>  [6]  -0.41629460  -0.26890732 -11.61569597  -0.02645597  -0.34758937

pcondigl21(v, u, alpha = 6)

#>  [1] 0.367517415 0.202267370 0.825425218 0.000291085 0.999961459 0.001544306
#>  [7] 0.001353188 1.000000000 0.141933361 0.999998899

qcondigl21(v, u, alpha = 6)

#>  [1] 0.1204253 0.1584193 0.6289526 0.8898574 0.3855237 0.8922194 0.9535514
#>  [8] 0.8484575 0.5537787 0.3055268

pcondigl12(u, v, alpha = 6)

#>  [1]  5.529694e-01  7.632423e-01  2.905738e-01  9.999573e-01  1.392047e-04
#>  [6]  9.997266e-01  9.997780e-01 -2.220446e-16  9.216055e-01  3.539984e-06

qcondigl12(u, v, alpha = 6)

#>  [1] 0.1253460 0.1627296 0.6762018 0.6168242 0.6525954 0.6750200 0.8416717
#>  [8] 0.9923653 0.4208920 0.5751476

rigl(10, alpha = 3)

#> # A tibble: 10 x 2
#>         u      v
#>     <dbl>  <dbl>
#>  1 0.935  0.925 
#>  2 0.212  0.330 
#>  3 0.652  0.608 
#>  4 0.126  0.0776
#>  5 0.267  0.524 
#>  6 0.386  0.485 
#>  7 0.0134 0.367 
#>  8 0.382  0.0930
#>  9 0.870  0.896 
#> 10 0.340  0.285